float -> double

+++ src/main/java/de/lergin/sponge/jobs/data/JobKeys.java
@@ -9,6 +9,6 @@ import static org.spongepowered.api.data.DataQuery.of;
 import static org.spongepowered.api.data.key.KeyFactory.makeSingleKey;
 
 public class JobKeys {
-    public static final Key<MapValue<String, Float>> JOB_DATA = makeMapKey(String.class, Float.class, of("JobData"));
+    public static final Key<MapValue<String, Double>> JOB_DATA = makeMapKey(String.class, Double.class, of("JobData"));
     public static final Key<Value<Boolean>> JOB_ENABLED = makeSingleKey(Boolean.class, Value.class, of("JobEnabled"));
 }
